+++ Playground/Playground.Wasm/Extensions/IntExtensions.cs
@@ -5,10 +5,10 @@ namespace Playground.Wasm.Extensions
 {
     public static class IntExtensions
     {
-        public static MudColor ToMudColor(this int argb) => new(
-            ColorUtils.RedFromArgb(argb),
-            ColorUtils.GreenFromArgb(argb),
-            ColorUtils.BlueFromArgb(argb),
-            ColorUtils.AlphaFromArgb(argb));
+        public static MudColor ToMudColor(this uint argb) => new(
+            (int)ColorUtils.RedFromArgb(argb),
+            (int)ColorUtils.GreenFromArgb(argb),
+            (int)ColorUtils.BlueFromArgb(argb),
+            (int)ColorUtils.AlphaFromArgb(argb));
     }
 }
